WebSockets/tests/submissions/Opera/cookies/006.html

<!doctype html>
<title>WebSockets: setting Secure cookie with document.cookie, checking ws request</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
async_test(function(t) {
  if (window.WebSocket) {
    document.cookie = 'ws_test=test; Path=/; Secure';
  }
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/ws/echo-cookie');
  ws.onmessage = t.step_func(function(e) {  
    ws.onclose = debug;
    ws.close();
    if (SCHEME_DOMAIN_PORT.substr(0,3) == 'wss') {
      assert_regexp_matches(e.data, /ws_test=test/);
    } else {
      assert_false(/ws_test=test/.test(e.data));
    }
    t.done();
  })
  ws.onclose = t.step_func(function() {assert_unreached()});
});
// remove cookie
document.cookie = 'ws_test=; Path=/; Secure; Expires=Sun, 06 Nov 1994 08:49:37 GMT';
</script>
